+++ tools/azure-sdk-tools/tests/test_pyproject_interactions.py
@@ -4,6 +4,7 @@ import os, tempfile, shutil
 import pytest
 
 from ci_tools.parsing import update_build_config, get_build_config, get_config_setting
+from ci_tools.environment_exclusions import is_check_enabled
 
 integration_folder = os.path.join(os.path.dirname(__file__), "integration")
 pyproject_folder = os.path.join(integration_folder, "scenarios", "sample_pyprojects")
@@ -57,6 +58,23 @@ def test_nonpresent_pyproject_update():
 
         reloaded_build_config = get_build_config(new_path)
         assert reloaded_build_config == update_result
+
+@pytest.mark.parametrize(
+    "check_name, environment_value, expected_result",
+    [
+        # check is true
+        ("mindependency", "true", True),
+        ("mindependency", "false", False)
+    ]
+)
+def test_environment_override(check_name, environment_value, expected_result):
+    with tempfile.TemporaryDirectory() as temp_dir:
+        new_path = shutil.copy(pyproject_file, temp_dir)
+
+        os.environ[f"{os.path.basename(temp_dir).upper()}_{check_name.upper()}"] = environment_value
+        result = is_check_enabled(temp_dir, check_name)
+
+        assert result == expected_result
 
 
 def test_pyproject_update_check_override():
